files

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@
 from typing import Union
 
 
-def sum_mixed_list(mxd_lst: Union[float, int]) -> float:
+def sum_mixed_list(mxd_lst: Union[int, float]) -> float:
     ''' type-annotated function sum_mixed_list which takes
     a list mxd_lst of integers and floats and returns their sum as a float.'''
     sum = 0
